Move non-CLI code into modules
