+++ src/main.rs
@@ -1,68 +1,8 @@
-use std::fmt;
+mod databases;
+
+use databases::{run_query, DBType};
 use std::path::PathBuf;
-use std::vec::Vec;
-
-use clap::{arg, command, value_parser, Command, ValueEnum};
-use rusqlite;
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, ValueEnum)]
-enum DBType {
-    Sqlite,
-    Duckdb,
-}
-
-struct QueryResult {
-    fields: Vec<String>,
-    rows: Vec<Vec<String>>,
-}
-
-impl fmt::Display for DBType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-
-fn run_sqlite_query(path: &PathBuf, query: &str) -> Result<QueryResult, rusqlite::Error> {
-    let conn = rusqlite::Connection::open(path)?;
-    let mut prepared = conn.prepare(query)?;
-    let num_columns = prepared.column_count();
-    let mut fields: Vec<String> = Vec::new();
-    for column_index in 0..num_columns {
-        fields.push(String::from(prepared.column_name(column_index)?))
-    }
-
-    let mut rows = prepared.query([])?;
-    let mut results: Vec<Vec<String>> = Vec::new();
-    while let Some(row) = rows.next()? {
-        let mut result: Vec<String> = Vec::new();
-        for column_index in 0..num_columns {
-            result.push(row.get(column_index)?);
-        }
-        results.push(result);
-    }
-    Ok(QueryResult {
-        fields,
-        rows: results,
-    })
-}
-
-fn run_query(path: &PathBuf, query: &str, db_type: &DBType) -> Result<(), &'static str> {
-    match db_type {
-        DBType::Sqlite => {
-            match run_sqlite_query(path, query) {
-                Ok(result) => {
-                    println!("Result schema: {:#?}", result.fields);
-                    println!("Results: {:#?}", result.rows);
-                }
-                Err(err) => {
-                    println!("SQLite error: {}", err);
-                }
-            }
-            Ok(())
-        }
-        _ => Err("Unsupported DB type"),
-    }
-}
+use clap::{arg, command, value_parser, Command};
 
 fn main() -> Result<(), &'static str> {
     let matches = command!()
